+++ src/motion/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_library(motion SHARED motion.c motion.h)
+add_library(motion SHARED motion.c)
 target_link_libraries(motion m)
