+++ tst/test_librange.c
@@ -2,7 +2,7 @@
 #include <math.h>
 #include <stdio.h>
 
-#include "range.h"
+#include "porthos/range.h"
 
 #define CALL(x) ck_assert_msg((0 == (x)), "Should succeed");
 #define ck_assert_dbl_eq_msg(X, Y, msg) ck_assert_msg((fabs((X)-(Y)) < 1e-6), msg);
